+++ crates/era-utils/tests/it/history.rs
@@ -1,11 +1,10 @@
-use alloy_primitives::bytes::Bytes;
-use futures_util::{Stream, TryStreamExt};
-use reqwest::{Client, IntoUrl, Url};
+use crate::ClientWithFakeIndex;
+use reqwest::{Client, Url};
 use reth_db_common::init::init_genesis;
-use reth_era_downloader::{EraClient, EraStream, EraStreamConfig, HttpClient};
+use reth_era_downloader::{EraClient, EraStream, EraStreamConfig};
 use reth_etl::Collector;
 use reth_provider::test_utils::create_test_provider_factory;
-use std::{future::Future, str::FromStr};
+use std::str::FromStr;
 use tempfile::tempdir;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
@@ -34,39 +33,4 @@ async fn test_history_imports_from_fresh_state_successfully() {
     let actual_block_number = reth_era_utils::import(stream, &pf, &mut hash_collector).unwrap();
 
     assert_eq!(actual_block_number, expected_block_number);
-}
-
-/// An HTTP client pre-programmed with canned answer to index.
-///
-/// Passes any other calls to a real HTTP client!
-#[derive(Debug, Clone)]
-struct ClientWithFakeIndex(Client);
-
-impl HttpClient for ClientWithFakeIndex {
-    fn get<U: IntoUrl + Send + Sync>(
-        &self,
-        url: U,
-    ) -> impl Future<
-        Output = eyre::Result<impl Stream<Item = eyre::Result<Bytes>> + Send + Sync + Unpin>,
-    > + Send
-           + Sync {
-        let url = url.into_url().unwrap();
-
-        async move {
-            match url.to_string().as_str() {
-                "https://era.ithaca.xyz/era1/index.html" => {
-                    Ok(Box::new(futures::stream::once(Box::pin(async move {
-                        Ok(bytes::Bytes::from_static(b"<a href=\"https://era.ithaca.xyz/era1/mainnet-00000-5ec1ffb8.era1\">mainnet-00000-5ec1ffb8.era1</a>"))
-                    })))
-                        as Box<dyn Stream<Item = eyre::Result<Bytes>> + Send + Sync + Unpin>)
-                }
-                _ => {
-                    let response = Client::get(&self.0, url).send().await?;
-
-                    Ok(Box::new(response.bytes_stream().map_err(|e| eyre::Error::new(e)))
-                        as Box<dyn Stream<Item = eyre::Result<Bytes>> + Send + Sync + Unpin>)
-                }
-            }
-        }
-    }
 }
